# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "User segmentation service",
        "title": "User Segmentation API",
        "contact": {
            "name": "Balin Maxim",
            "email": "linqcod@yandex.ru"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/segments": {
            "post": {
                "description": "create segment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segments"
                ],
                "summary": "create segment",
                "parameters": [
                    {
                        "description": "Create segment",
                        "name": "segment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_linqcod_avito-internship-2023_internal_model.CreateSegmentDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "segment created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_linqcod_avito-internship-2023_internal_model.CreateSegmentResponse"
                        }
                    },
                    "400": {
                        "description": "error bad request data",
                        "schema": {
                            "$ref": "#/definitions/github_com_linqcod_avito-internship-2023_internal_model.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "error while inserting segment to db table",
                        "schema": {
                            "$ref": "#/definitions/github_com_linqcod_avito-internship-2023_internal_model.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/segments/{id}": {
            "delete": {
                "description": "delete segment by id",
                "tags": [
                    "segments"
                ],
                "summary": "delete segment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Segment slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "segment deleted successfully"
                    },
                    "500": {
                        "description": "error while deleting segment",
                        "schema": {
                            "$ref": "#/definitions/github_com_linqcod_avito-internship-2023_internal_model.ErrorDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_linqcod_avito-internship-2023_internal_model.CreateSegmentDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "github_com_linqcod_avito-internship-2023_internal_model.CreateSegmentResponse": {
            "type": "object",
            "properties": {
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "github_com_linqcod_avito-internship-2023_internal_model.ErrorDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}